+++ js/map.js
@@ -20,27 +20,13 @@ const createMap = (lat, lng, scale, selector) => {
   return mapCanvas;
 };
 
-const map = createMap(35.652832, 139.839478, 12, 'map-canvas');
-
 const createPinIcon = (url, size, anchor) => L.icon({
   iconUrl: url,
   iconSize: size,
   iconAnchor: anchor,
 });
 
-const mainPinIcon = createPinIcon(
-  './img/main-pin.svg',
-  [52, 52],
-  [26, 52],
-);
-
-const regularIcon = createPinIcon(
-  './img/pin.svg',
-  [40, 40],
-  [20, 40],
-);
-
-const createPin = (lat, lng, drg, icn) => {
+const createPin = (lat, lng, drg, icn, map) => {
   const pin = L.marker(
     {
       lat: lat,
@@ -55,23 +41,14 @@ const createPin = (lat, lng, drg, icn) => {
   return pin;
 };
 
-const mainPin = createPin(
-  35.652832,
-  139.839478,
-  true,
-  mainPinIcon);
-
-const address = document.querySelector('#address');
-
-mainPin.on('moveend', (evt) => {
-  address.value = `${evt.target._latlng.lat.toFixed(5)}, ${evt.target._latlng.lng.toFixed(5)}`;
-});
-
-//
-
-const createPinBaloon = (lat, lng, baloonCntnt) => {
-  const p = createPin(lat, lng, false, regularIcon);
+const createPinBaloon = (lat, lng, baloonCntnt, icn, map) => {
+  const p = createPin(lat, lng, false, icn, map);
   p.bindPopup(baloonCntnt);
 };
 
-export {createPinBaloon};
+export {
+  createPinBaloon,
+  createMap,
+  createPin,
+  createPinIcon
+};
